ChatbotMigration/requirements/requirements.txt
# Python Dependencies for Offline Retrieval-Based Chatbot
# Install using: pip install -r requirements.txt

# Core ML/NLP Libraries
sentence-transformers==2.2.2
transformers==4.33.0
torch==2.0.1

# Vector Search & Similarity
faiss-cpu==1.7.4
scikit-learn==1.3.0

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Web Framework (if using Streamlit)
streamlit==1.28.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Optional: For advanced features
nltk==3.8.1
spacy==3.6.1
